requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7

